refactor: Minor bug fix #1
